Refactor ArchivedMessages to clean architecture
