+++ source/ArchivedMessages.Infrastructure/QueryBuilder.cs
@@ -13,7 +13,7 @@
 // limitations under the License.
 
 using Dapper;
-using Energinet.DataHub.EDI.ArchivedMessages.Interfaces;
+using Energinet.DataHub.EDI.ArchivedMessages.Interfaces.Models;
 using Energinet.DataHub.EDI.BuildingBlocks.Domain.Authentication;
 using Energinet.DataHub.EDI.BuildingBlocks.Domain.Exceptions;
