+++ source/ArchivedMessages.Interfaces/ArchivedMessages.Interfaces.csproj
@@ -1,4 +1,4 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
     <AssemblyName>Energinet.DataHub.EDI.ArchivedMessages.Interfaces</AssemblyName>
     <RootNamespace>Energinet.DataHub.EDI.ArchivedMessages.Interfaces</RootNamespace>
@@ -15,7 +15,7 @@
   </ItemGroup>
   <!--Giving the integration test project of ArchivedMessages access to internals -->
   <ItemGroup>
-    <InternalsVisibleTo Include="Energinet.DataHub.EDI.ArchivedMessages.IntegrationTests"/>
+    <InternalsVisibleTo Include="Energinet.DataHub.EDI.ArchivedMessages.IntegrationTests" />
   </ItemGroup>
   <ItemDefinitionGroup>
     <ProjectReference>
