+++ source/ArchivedMessages.Interfaces/IArchivedMessageRepository.cs
@@ -12,9 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Energinet.DataHub.EDI.ArchivedMessages.Interfaces;
+using Energinet.DataHub.EDI.ArchivedMessages.Interfaces.Models;
 
-namespace Energinet.DataHub.EDI.ArchivedMessages.Infrastructure;
+namespace Energinet.DataHub.EDI.ArchivedMessages.Interfaces;
 
 /// <summary>
 /// Responsible for archiving messages.
