+++ source/ArchivedMessages.Interfaces/Models/FieldToSortBy.cs
@@ -12,14 +12,17 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Energinet.DataHub.EDI.ArchivedMessages.Interfaces;
+namespace Energinet.DataHub.EDI.ArchivedMessages.Interfaces.Models;
 
-public readonly struct DirectionToSortBy
+public readonly struct FieldToSortBy
 {
-    public static readonly DirectionToSortBy Ascending = new("ASC");
-    public static readonly DirectionToSortBy Descending = new("DESC");
+    public static readonly FieldToSortBy MessageId = new("MessageId");
+    public static readonly FieldToSortBy DocumentType = new("DocumentType");
+    public static readonly FieldToSortBy SenderNumber = new("SenderNumber");
+    public static readonly FieldToSortBy ReceiverNumber = new("ReceiverNumber");
+    public static readonly FieldToSortBy CreatedAt = new("CreatedAt");
 
-    private DirectionToSortBy(string identifier)
+    private FieldToSortBy(string identifier)
     {
         Identifier = identifier;
     }
