+++ source/B2CWebApi/Mappers/FieldToSortByMapper.cs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Energinet.DataHub.EDI.ArchivedMessages.Interfaces;
+using Energinet.DataHub.EDI.ArchivedMessages.Interfaces.Models;
 
 namespace Energinet.DataHub.EDI.B2CWebApi.Mappers;
