+++ source/OutgoingMessages.IntegrationTests/OutgoingMessages.IntegrationTests.csproj
@@ -33,7 +33,7 @@
     </ItemGroup>
 
     <ItemGroup>
-        <ProjectReference Include="..\ArchivedMessages.Application\ArchivedMessages.Application.csproj" />
+        <ProjectReference Include="..\ArchivedMessages.Infrastructure\ArchivedMessages.Infrastructure.csproj" />
         <ProjectReference Include="..\B2BApi\B2BApi.csproj" />
         <ProjectReference Include="..\BuildingBlocks.Tests\BuildingBlocks.Tests.csproj" />
         <ProjectReference Include="..\BuildingBlocks.Infrastructure\BuildingBlocks.Infrastructure.csproj" />
